Рефакторинг. Разбивает app.exception-filter
