+++ src/shared/libs/rest/exception-filter/app.exception-filter.ts
@@ -1,0 +1,24 @@
+import { Request, Response, NextFunction } from 'express';
+import { StatusCodes } from 'http-status-codes';
+import { injectable, inject } from 'inversify';
+import { Component } from '../../../const.js';
+import { createErrorObject } from '../../../helpers/common.js';
+import { ApplicationError } from '../types/application-error.enum.js';
+import { ExceptionFilter } from './exception-filter.interface.js';
+import { Logger } from '../../logger/logger.interface.js';
+
+@injectable()
+export class AppExceptionFilter implements ExceptionFilter {
+  constructor(
+    @inject(Component.Logger) private readonly logger: Logger
+  ) {
+    this.logger.info('Register AppExceptionFilter');
+  }
+
+  public catch(error: Error, _req: Request, res: Response, _next: NextFunction): void {
+    this.logger.error(error.message, error);
+    res
+      .status(StatusCodes.INTERNAL_SERVER_ERROR)
+      .json(createErrorObject(ApplicationError.ServiceError, error.message));
+  }
+}
