+++ src/shared/libs/rest/exception-filter/http-error.exception-filter.ts
@@ -1,0 +1,30 @@
+import { Request, Response, NextFunction } from 'express';
+import { StatusCodes } from 'http-status-codes';
+import { injectable, inject } from 'inversify';
+import { Component } from '../../../const.js';
+import { createErrorObject } from '../../../helpers/common.js';
+import { HttpError } from '../errors/http.error.js';
+import { ApplicationError } from '../types/application-error.enum.js';
+import { ExceptionFilter } from './exception-filter.interface.js';
+import { Logger } from '../../logger/logger.interface.js';
+
+@injectable()
+export class HttpErrorExceptionFilter implements ExceptionFilter {
+  constructor(
+    @inject(Component.Logger) private readonly logger: Logger
+  ) {
+    this.logger.info('Register HttpErrorExceptionFilter');
+  }
+
+  public catch(error: unknown, req: Request, res: Response, next: NextFunction): void {
+    if (!(error instanceof HttpError)) {
+      return next(error);
+    }
+
+    this.logger.error(`[HttpErrorException]: ${req.path} # ${error.message}`, error);
+
+    res
+      .status(StatusCodes.BAD_REQUEST)
+      .json(createErrorObject(ApplicationError.CommonError, error.message));
+  }
+}
